new function for selecting best loci based on coverage

+++ secapr/reference_assembly.py
@@ -297,73 +297,155 @@ def clean_with_picard(sample_output_folder,sample_id,sorted_bam,log):
 	return picard_out, dupl_out
 
 
-def bam_consensus(reference,bam_file,name_base,out_dir,min_cov):
+def bcf_cons(pileup,out_fasta,cov):
+	with open(pileup) as f:
+	    content = [x.strip('\n') for x in f.readlines()]
+
+	loci = []
+	seq_dict = {}
+	for line in content:
+	    if '#' not in line:
+	        # Split the tab delimited lines into their segments
+	        element = line.split('\t')
+	        # Create a list of all loci names
+	        seq_name = element[0]
+	        if seq_name not in loci:
+	            loci.append(seq_name)
+	        # By default call every position a uncertainty
+	        basecall = "N"
+
+	        # Turn all lower case values in upper case
+	        sample = element[4].upper()
+	        # make a directory with all different basecalls and count their occurences
+	        calls = dict((letter,sample.count(letter)) for letter in set(sample))
+
+	        # The basecall in the reference
+	        reference = element[2]
+	        # These characters signal a match with the reference
+	        match_ref = "." ","
+	        # List of base characters
+	        bases = "A" "G" "C" "T"
+
+	        # find out how many agreements with reference are among the basecalls. These are all . and , basecalls listed in match_ref.
+	        # reset the counter before every round (every line in file aka position in sequence)
+	        list_matches = 0
+	        for key,value in calls.items():
+	            if key in match_ref:
+	                list_matches += value
+	        if list_matches >= cov:
+	            basecall = reference
+
+	        # find if there are any well supported SNPs and make the most prominent call
+	        for key in sorted(calls, key=calls.get, reverse=True):
+	            if key in bases:
+	                if int(calls[key]) >= cov:
+	                    if int(calls[key]) >= list_matches:
+	                        basecall = key
+	                        break
+	        # add the final basecall to the dictionary and to the respective key if it already exists, otherwise create new key
+	        seq_dict.setdefault(element[0],[])
+	        seq_dict[element[0]].append(basecall)
+	# Join all basecalls for each key (=locus) into one sequence and deposit in new dictionary
+	concat_basecalls = {}
+	for key, value in seq_dict.items():
+		concat_basecalls[key] = "".join(value)
+	#print concat_basecalls
+
+	with open(out_fasta, "wb") as f:
+		for k, v in concat_basecalls.items():
+			f.write(">" + k+ "\n")
+			f.write(v+ "\n")
+	return out_fasta
+
+
+def bam_consensus(reference,bam_file,name_base,out_dir,min_cov,cons_method = 'custom'):
 	print ("Generating a consensus sequence from bam-file..........")
 	# Creating consensus sequences from bam-files
 
-	# 1. mpilup__________
-	mpileup = [
-		"samtools",
-		"mpileup",
-		"-u",
-		"-f",
-		reference,
-		bam_file
-	]
-	mpileup_file = os.path.join(out_dir, "%s.mpileup" %name_base)
-	with open(mpileup_file, 'w') as mpileupfile:
-		mp = subprocess.Popen(mpileup, stdout=mpileupfile, stderr=subprocess.PIPE)
-		mp.communicate()
-		mp.wait()
-
-	# 2. bcftools__________
-	vcf_file = os.path.join(out_dir, "%s.vcf" %name_base)
-	#bcf_cmd = [
-	#	"bcftools",
-	#	"view",
-	#	"-c",
-	#	"-g",
-	#	mpileup_file
-	#]
-	bcf_cmd = [
-		"bcftools",
-		"call",
-		"-c",
-		mpileup_file
-	]
-	with open(vcf_file, 'w') as vcffile:
-		vcf = subprocess.Popen(bcf_cmd, stdout=vcffile, stderr=subprocess.PIPE)
-		vcf.communicate()
-		vcf.wait()
-
-	# 3. vcfutils__________
-	fq_file = os.path.join(out_dir, "%s.fq" %name_base)
-	vcfutils_cmd = [
-		"vcfutils.pl",
-		"vcf2fq",
-		"-d",
-		str(min_cov),
-		vcf_file
-	]
-	with open(fq_file, 'w') as fqfile:
-		fq = subprocess.Popen(vcfutils_cmd, stdout=fqfile)
-		fq.communicate()
-		fq.wait()
+	# This is the custom written consensus building function
+	if cons_method == 'custom':
+		mpileup = [
+			"samtools",
+			 "mpileup",
+			"-A",
+			"-f",
+			reference,
+			bam_file
+		]
+		mpileup_file = os.path.join(out_dir, "%s.bcf" %name_base)
+		with open(mpileup_file, 'w') as mpileupfile:
+			mp = subprocess.Popen(mpileup, stdout=mpileupfile, stderr=subprocess.PIPE)
+			mp.communicate()
+			mp.wait()
+		fasta_file = os.path.join(out_dir,"%s_temp.fasta" %name_base)
+		fasta_file = bcf_cons(mpileup_file,fasta_file,min_cov)
 	
-	# 4. seqtk__________
-	# Converting fq into fasta files
-	fasta_file = os.path.join(out_dir,"%s_temp.fasta" %name_base)
-	make_fasta = [
-		"seqtk",
-		"seq",
-		"-a",
-		fq_file,
-	]
-	with open(fasta_file, 'w') as fastafile:
-		fasta = subprocess.Popen(make_fasta, stdout=fastafile)
-		fasta.communicate()
-		fasta.wait()
-	
+	# This is the bcftools version which sometimes acts a bit funny
+	else:
+		# 1. mpilup__________
+		mpileup = [
+			"samtools",
+			"mpileup",
+			"-u",
+			"-f",
+			reference,
+			bam_file
+		]
+		mpileup_file = os.path.join(out_dir, "%s.mpileup" %name_base)
+		with open(mpileup_file, 'w') as mpileupfile:
+			mp = subprocess.Popen(mpileup, stdout=mpileupfile, stderr=subprocess.PIPE)
+			mp.communicate()
+			mp.wait()
+
+		# 2. bcftools__________
+		vcf_file = os.path.join(out_dir, "%s.vcf" %name_base)
+		#bcf_cmd = [
+		#	"bcftools",
+		#	"view",
+		#	"-c",
+		#	"-g",
+		#	mpileup_file
+		#]
+		bcf_cmd = [
+			"bcftools",
+			"call",
+			"-c",
+			mpileup_file
+		]
+		with open(vcf_file, 'w') as vcffile:
+			vcf = subprocess.Popen(bcf_cmd, stdout=vcffile, stderr=subprocess.PIPE)
+			vcf.communicate()
+			vcf.wait()
+
+		# 3. vcfutils__________
+		fq_file = os.path.join(out_dir, "%s.fq" %name_base)
+		vcfutils_cmd = [
+			"vcfutils.pl",
+			"vcf2fq",
+			"-d",
+			str(min_cov),
+			vcf_file
+		]
+		with open(fq_file, 'w') as fqfile:
+			fq = subprocess.Popen(vcfutils_cmd, stdout=fqfile)
+			fq.communicate()
+			fq.wait()
+		
+		# 4. seqtk__________
+		# Converting fq into fasta files
+		fasta_file = os.path.join(out_dir,"%s_temp.fasta" %name_base)
+		make_fasta = [
+			"seqtk",
+			"seq",
+			"-a",
+			fq_file,
+		]
+		with open(fasta_file, 'w') as fastafile:
+			fasta = subprocess.Popen(make_fasta, stdout=fastafile)
+			fasta.communicate()
+			fasta.wait()
+
+	# Independently of the chosen consensus building method, the following brings the resulting fasta file into the correct format for further processing
 	# Create a new fasta file for final fasta printing
 	final_fasta_file = os.path.join(out_dir,"%s.fasta" %name_base)
 
@@ -405,13 +487,19 @@ def bam_consensus(reference,bam_file,name_base,out_dir,min_cov):
 		tmp_folder = "%s/tmp" %out_dir
 		if not os.path.exists(tmp_folder):
 			os.makedirs(tmp_folder)
-		mpileup = glob.glob('%s/*.mpileup' %out_dir)[0]
-		vcf = glob.glob('%s/*.vcf' %out_dir)[0]
-		shutil.move(vcf,tmp_folder)
-		shutil.move(mpileup,tmp_folder)
+
+		if cons_method == 'custom':
+			bcf = glob.glob('%s/*.bcf' %out_dir)[0]
+			shutil.move(bcf,tmp_folder)
+		else:
+			mpileup = glob.glob('%s/*.mpileup' %out_dir)[0]
+			vcf = glob.glob('%s/*.vcf' %out_dir)[0]
+			shutil.move(vcf,tmp_folder)
+			shutil.move(mpileup,tmp_folder)
 
 	os.remove(fasta_file)
-	os.remove(fq_file)
+	if not cons_method == 'custom':
+		os.remove(fq_file)
 
 	return final_fasta_file
